Provide rewind

+++ libs/chan/reader.cc
@@ -153,14 +153,14 @@ uint64_t Reader::read_nbo_64() throw (Error)
 
 //--------------------------------------------------------------------------
 // Skip N bytes
-void Reader::skip(int n) throw (Error)
+void Reader::skip(size_t n) throw (Error)
 {
   read(0, n);
 }
 
 //--------------------------------------------------------------------------
 // Skip to given alignment (bytes) from current offset
-void Reader::align(int n) throw (Error)
+void Reader::align(size_t n) throw (Error)
 { 
   skip((size_t)(n*((offset+n-1)/n) - offset));  // Bytes to skip
 }
